refactor(lib::track): move metadata fetching to own module
