+++ lib/src/track/mod.rs
@@ -1,8 +1,9 @@
 mod fmt_ext;
 mod full_track;
+mod read_metadata;
 
 pub use fmt_ext::DurationFmtShort;
 pub use full_track::{
-    LyricData, MediaTypes, MediaTypesSimple, MetadataOptions, PodcastTrackData, RadioTrackData,
-    Track, TrackData, TrackMetadata, parse_metadata_from_file,
+    LyricData, MediaTypes, MediaTypesSimple, PodcastTrackData, RadioTrackData, Track, TrackData,
 };
+pub use read_metadata::{MetadataOptions, TrackMetadata, parse_metadata_from_file};
